Adicionado serviço para comentario e montador de postagens.

+++ src/java/modelo/dao/api/PostagemDAO.java
@@ -6,6 +6,6 @@ import modelo.Postagem;
 public interface PostagemDAO {
     
     public int inserir(Postagem postagem);
-    public Postagem encontrarPorIdentificador(Long id);
+    public Postagem encontrarPorIdentificador(Integer id);
     public List<Postagem> encontrarCincoPostagens();
 }
